WIP Start building main, data, models pipeline
